feat(origin-ui-core): improvements to user profile page
