+++ packages/origin-ui-core/src/styles/UserProfile.scss
@@ -1,0 +1,15 @@
+#user_profile {
+  .container {
+    padding: 20px;
+  }
+
+  .buttonAndIconHolder {
+    padding-top: 1rem;
+    display: flex;
+    align-items: center;
+  }
+
+  .infoIcon {
+    margin-left: 1rem;
+  }
+}
