+++ packages/origin-ui-core/src/styles/app.scss
@@ -4,3 +4,4 @@
 @import './FiltersHeader.scss';
 @import './SmartMeterReadingsChart.scss';
 @import './LoginPage.scss';
+@import './UserProfile.scss';
